requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cffi==1.15.1
cssselect2==0.7.0
Django==4.2.5
fonttools==4.42.1
html5lib==1.1
Pillow==10.0.0
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
